changed trans to int

+++ sparsematrix.c
@@ -40,10 +40,10 @@ int findIndex(int index, SparseRow* indexes, SparseRow* values) {
 }
 
 //get the transpose of the sparse matrix as a single sparse row
-uint8_t* transposeSparseMatrix(SparseRow* sparseValues, SparseRow* sparseIndexes) {
+int* transposeSparseMatrix(SparseRow* sparseValues, SparseRow* sparseIndexes) {
 
     // init all rows in the transpose matrix
-    uint8_t* transpose = calloc(DEFAULT_SIZE * DEFAULT_SIZE, sizeof(uint8_t));
+    int* transpose = calloc(DEFAULT_SIZE * DEFAULT_SIZE, sizeof(int));
     for(int i = 0; i < DEFAULT_SIZE; i++) {
         SparseRow row = sparseIndexes[i];
         for(int j = 0; j < row.size; j++) {
@@ -51,7 +51,7 @@ uint8_t* transposeSparseMatrix(SparseRow* sparseValues, SparseRow* sparseIndexes
             transpose[index * DEFAULT_SIZE + i] = sparseValues[i].col[j];
         }
     }
-    printf("\t\tAllocated %lu bytes for tranposed sparse matrix...\n", DEFAULT_SIZE * DEFAULT_SIZE * sizeof(uint8_t));
+    printf("\t\tAllocated %lu bytes for tranposed sparse matrix...\n", DEFAULT_SIZE * DEFAULT_SIZE * sizeof(int));
     return transpose;
 }
 
@@ -59,7 +59,7 @@ int* multiplySparseMatrices(MultiMatrix A, MultiMatrix B) {
     int* result = malloc(DEFAULT_SIZE * DEFAULT_SIZE * sizeof(int));
     printf("\t\tAllocated %lu bytes for sparse multiplication...\n", DEFAULT_SIZE * DEFAULT_SIZE * sizeof(int));
 
-    uint8_t* transpose = transposeSparseMatrix(B.values, B.indexes);
+    int* transpose = transposeSparseMatrix(B.values, B.indexes);
 
     int i, j, k, tmp;
     #pragma omp parallel for private(i, j, k) schedule(static, BLOCK_SIZE) reduction(+:tmp) //schedule(dynamic, 1)
